requirements.txt
fastapi
uvicorn
sqlmodel
python-dotenv
passlib[bcrypt]
pyjwt
